// src/app/pages/formulrio/interfaces/formularioResponse.interface.ts
export interface FormuarioInterfaceData {

    fecha:           string;
    nombre:          string;
    cultivo:         string;
    alcaldia:        string;
    pueblo:          string;
    paraje:          string;
    superficie:      number;
    numeroApliacion: string;
    folio:           string;
    suelo:           string;
    firma:           string;

}